// src/omt/schema.ts
export const OMT_SCHEMA: Readonly<Record<string, OmtLayer>> = {
	aerodrome_label: {
		minzoom: 8,
		maxzoom: 14,
		fields: [
			'class',
			'ele',
			'ele_ft',
			'iata',
			'icao',
			'name',
			'name:af',
			'name:am',
			'name:ar',
			'name:az',
			'name:be',
			'name:bg',
			'name:bn',
			'name:br',
			'name:bs',
			'name:ca',
			'name:co',
			'name:cs',
			'name:cy',
			'name:da',
			'name:de',
			'name:el',
			'name:en',
			'name:eo',
			'name:es',
			'name:et',
			'name:eu',
			'name:fa',
			'name:fi',
			'name:fr',
			'name:fy',
			'name:ga',
			'name:gd',
			'name:he',
			'name:hi',
			'name:hr',
			'name:hu',
			'name:hy',
			'name:id',
			'name:is',
			'name:it',
			'name:ja',
			'name:ja-Hira',
			'name:ja-Latn',
			'name:ka',
			'name:kk',
			'name:kn',
			'name:ko',
			'name:ko-Latn',
			'name:ku',
			'name:la',
			'name:latin',
			'name:lb',
			'name:lt',
			'name:lv',
			'name:mk',
			'name:ml',
			'name:mt',
			'name:nl',
			'name:no',
			'name:nonlatin',
			'name:oc',
			'name:pa',
			'name:pl',
			'name:pnb',
			'name:pt',
			'name:rm',
			'name:ro',
			'name:ru',
			'name:sk',
			'name:sl',
			'name:sq',
			'name:sr',
			'name:sr-Latn',
			'name:sv',
			'name:ta',
			'name:te',
			'name:th',
			'name:tr',
			'name:uk',
			'name:ur',
			'name:vi',
			'name:zh',
			'name:zh-Hans',
			'name:zh-Hant',
			'name_de',
			'name_en',
			'name_int',
		],
	},
// ...
};