// src/protomaps/schema.ts
export const PROTOMAPS_SCHEMA: Readonly<Record<string, ProtomapsLayer>> = {
	boundaries: { minzoom: 0, maxzoom: 15, fields: ['brk_a3', 'disputed', 'kind', 'kind_detail', 'sort_rank'] },
	buildings: {
		minzoom: 11,
		maxzoom: 15,
		fields: ['addr_housenumber', 'height', 'kind', 'kind_detail', 'layer', 'min_height', 'sort_rank'],
	},
	earth: {
		minzoom: 0,
		maxzoom: 15,
		fields: [
			'kind',
			'min_zoom',
			'name',
			'name2',
			'name3',
			'name:ar',
			'name:bg',
			'name:cs',
			'name:da',
			'name:de',
			'name:el',
			'name:en',
			'name:es',
			'name:et',
			'name:fa',
			'name:fi',
			'name:fr',
			'name:ga',
			'name:he',
			'name:hi',
			'name:hr',
			'name:hu',
			'name:id',
			'name:it',
			'name:ja',
			'name:ko',
			'name:lt',
			'name:lv',
			'name:mr',
			'name:mt',
			'name:ne',
			'name:nl',
			'name:no',
			'name:pl',
			'name:pt',
			'name:ro',
			'name:ru',
			'name:sk',
			'name:sl',
			'name:sv',
			'name:tr',
			'name:uk',
			'name:ur',
			'name:vi',
			'name:zh-Hans',
			'name:zh-Hant',
			'pgf:name:hi',
			'pgf:name:mr',
			'pgf:name:ne',
			'script',
			'script2',
			'script3',
		],
	},
// ...
};